+++ initiatory.go
@@ -38,6 +38,7 @@ func init() {
 
 const (
 	/* conversional */
+
 	// HTMLEngine conversion for config.HTMLEngine
 	HTMLEngine = config.HTMLEngine
 	// PongoEngine conversion for config.PongoEngine